+++ damt.ts
@@ -9,17 +9,18 @@
  *
  * @date originally created: 25 Feb 2022.
  * @date updated significantly: November 2022.
+ * @date significant refactoring: April 2024.
  *
  * @details Program is used to manage acronyms stored in a SQLite database.
  * Application is written in TypeScript for use with the Deno runtime: https://deno.land/
  *
  * @note The program can be run directly from the GitHub repo using the command:
  * @code deno run --quiet --allow-read --allow-env=ACRODB --allow-write --unstable-temporal https://raw.githubusercontent.com/wiremoons/damt/main/damt.ts
- * @note The program can be run with Deno using the command:
- * @code deno run --quiet --allow-read --allow-env=ACRODB --allow-write --unstable-temporal
- * @note The program can be installed to 'DENO_INSTALL_ROOT' to using the command:
+ * @note A local copy of the program `damt.ts` can be run with Deno using the command:
+ * @code deno run --quiet --allow-read --allow-env=ACRODB --allow-write --unstable-temporal damt.ts
+ * @note A local copy of the program `damt.ts` can be installed to 'DENO_INSTALL_ROOT' to using the command:
  * @code deno install -f --quiet --allow-read --allow-env=ACRODB --allow-write --unstable-temporal damt.ts
- * @note The program can be compiled using the command:
+ * @note A local copy of the program `damt.ts` can be compiled using the command:
  * @code deno compile --quiet --allow-read --allow-env=ACRODB --allow-write --unstable-temporal damt.ts
  */
 
@@ -28,28 +29,22 @@
 //--------------------------------
 
 // Deno stdlib imports
-import { parseArgs } from "https://deno.land/std@0.220.1/cli/parse_args.ts";
-import { format } from "https://deno.land/std@0.220.1/fmt/bytes.ts";
+import { parseArgs } from "jsr:@std/cli@0.223.0/parse-args";
+import { format } from "jsr:@std/fmt@0.223.0/bytes";
 import {
   basename,
   dirname,
   fromFileUrl,
   join,
   normalize,
-} from "https://deno.land/std@0.220.1/path/mod.ts";
-import {
-  //blue,
-  bold,
-  //underline,
-} from "https://deno.land/std@0.220.1/fmt/colors.ts";
+} from "jsr:@std/path@0.223.0";
+import { bold } from "jsr:@std/fmt@0.223.0/colors";
 
 // Other imports
 import {
   cliVersion,
   existsFile,
   getFileModTime,
-  //isNumber,
-  isString,
 } from "https://deno.land/x/deno_mod@0.8.1/mod.ts";
 import { DB } from "https://deno.land/x/sqlite@v3.8/mod.ts";
 
@@ -57,23 +52,17 @@ import { DB } from "https://deno.land/x/sqlite@v3.8/mod.ts";
 // GLOBAL DECLARATIONS
 //--------------------------------
 
-// DATABASE INTERFACE DECLARATION
-
-interface DamtInterface {
-  dbFileName: string;
-  dbFullPath: string;
-  dbSize?: string;
-  dbLastAccess?: string;
-  dbSqliteVersion?: string;
-  dbRecordCount?: string;
-  dbNewestAcronym?: string;
-}
-
 // CLI VERSION OPTIONS DECLARATION
 
-/** define options for `cliVersion()` function for application version data */
+/** define options for `cliVersion()` function for application version data
+ * @typedef versionOptions
+ * @prop {string} version the current version of this application
+ * @prop {string} copyrightName the name of the person that hold the copyright of this application
+ * @prop {string} licenseUrl the web site address that holds a copy of this applications license
+ * @prop {string} crYear the year(s) the copyright applies to for this application
+ */
 const versionOptions = {
-  version: "0.3.1",
+  version: "0.4.0",
   copyrightName: "Simon Rowe",
   licenseUrl: "https://github.com/wiremoons/damt/",
   crYear: "2022-2024",
@@ -92,27 +81,31 @@ const cliOpts = {
 //--------------------------------
 
 /** obtain any command line arguments and exec them as needed */
-async function execCliArgs(db: DB, dbData: DamtInterface) {
+async function execCliArgs(db: DB) {
   //console.log(parse(Deno.args,cliOpts));
   const cliArgs = parseArgs(Deno.args, cliOpts);
 
   if (cliArgs.search) {
-    dbSearch(Deno.args[1], db, dbData);
+    dbSearch(Deno.args[1], db);
+    db.close();
     Deno.exit(0);
   }
 
   if (cliArgs.latest) {
-    dbLatestRecords(db, dbData);
+    dbLatestRecords(db);
+    db.close();
     Deno.exit(0);
   }
 
   if (cliArgs.help) {
     printHelp();
+    db.close();
     Deno.exit(0);
   }
 
   if (cliArgs.version) {
     await printVersionInfo();
+    db.close();
     Deno.exit(0);
   }
 }
@@ -131,8 +124,17 @@ async function execCliArgs(db: DB, dbData: DamtInterface) {
 // UTILITY FUNCTIONS
 //--------------------------------
 
-/** Obtain `filePath` file size */
-async function getFileSize(filePath: string): Promise<number | undefined> {
+/** Return the name of the currently running program without the path included. */
+function getAppName(): string {
+  return `${basename(Deno.mainModule) ?? "UNKNOWN"}`;
+}
+
+/** Using the filePath to obtain database file size
+ * NB: `format` from "jsr:@std/fmt@0.223.0/bytes" to show DB size in bytes/MB/GB etc
+ */
+async function getFileSize(filePath: string): Promise<string> {
+  // check have valid filePath
+  if (typeof filePath !== "string") return "UNKNOWN";
   // check for URL path instead of OS path
   if (filePath.startsWith("file:")) {
     filePath = fromFileUrl(filePath);
@@ -141,19 +143,22 @@ async function getFileSize(filePath: string): Promise<number | undefined> {
   try {
     const fileInfo = await Deno.lstat(filePath);
     if (fileInfo.isFile) {
-      return fileInfo.size;
+      return format(fileInfo.size);
     }
-  } catch {
-    return undefined;
+  } catch (err) {
+    console.error(`ERROR: Failed to get DB file size: '${err}'`);
+    return "UNKNOWN";
   }
+  return "UNKNOWN";
 }
 
 /**
- * Type Guard for DamtInterface interface object
+ * Type Guards
  */
+/** Type guard for string */
 // deno-lint-ignore no-explicit-any
-function isObject(arg: any): arg is DamtInterface {
-  return arg !== undefined;
+export function isString(arg: any): arg is string {
+  return arg !== undefined && typeof arg === "string";
 }
 
 /** Convert epoch date to date and time for display in output as a string */
@@ -184,20 +189,8 @@ function getDisplayDateTime(epochTime: number): string {
 }
 
 //--------------------------------
-// APPLICATION FUNCTIONS
+// DATABASE LOCATE FUNCTIONS
 //--------------------------------
-
-/** Return the name of the currently running program without the path included. */
-function getAppName(): string {
-  return `${basename(Deno.mainModule) ?? "UNKNOWN"}`;
-}
-
-/** Check executable location for valid database file named: 'acronyms.db'  */
-async function getLocalDBFile(): Promise<string | undefined> {
-  const appDB = join(normalize(dirname(Deno.mainModule)), "acronyms.db");
-  console.log(`Constructed filename: ${appDB}`);
-  return await existsFile(appDB) ? appDB : undefined;
-}
 
 /** Check environment variable 'ACRODB' for a valid filename */
 async function getDBEnv(): Promise<string | undefined> {
@@ -205,63 +198,68 @@ async function getDBEnv(): Promise<string | undefined> {
   return await existsFile(dbFile) ? dbFile : undefined;
 }
 
-//** Populate the field in the 'DamtInterface' with the database information */
-async function setDbData(): Promise<DamtInterface> {
-  // check environment first and fallback to executable directory looking DB file location
+/** Check executable location for valid database file named: 'acronyms.db'  */
+async function getLocalDBFile(): Promise<string | undefined> {
+  const appDB = join(normalize(dirname(Deno.mainModule)), "acronyms.db");
+  // DEBUG OUTPUT:
+  //console.log(`Constructed filename: ${appDB}`);
+  return await existsFile(appDB) ? appDB : undefined;
+}
+
+/** Attempts to find the database file on the local file system */
+async function findDbFileLocation(): Promise<string> {
+  // check environment first and fallback to executable directory
   const dbLocation = await getDBEnv() || await getLocalDBFile();
   if (!dbLocation) {
     return Promise.reject(
-      new Error("failed to locate a valid database filename."),
+      new Error(
+        "'findDbFileLocation()' failed to locate a valid database filename.",
+      ),
     );
   }
-
-  // Use https://deno.land/std/fmt/ to show DB size in formatted bytes/MB/GB etc
-  const prettyDBSze = await getFileSize(dbLocation).then((size) =>
-    size ? format(size) : "UNKNOWN"
-  );
-
-  return {
-    dbFileName: basename(dbLocation),
-    dbFullPath: normalize(dbLocation),
-    dbSize: prettyDBSze,
-    dbLastAccess: await getFileModTime(dbLocation),
-  };
+  return dbLocation;
 }
 
-//** Try to open the database file and return the handle to it
-function openDB(dbData: DamtInterface): DB | undefined {
-  if ((isObject(dbData)) && (isString(dbData.dbFullPath))) {
-    return new DB(dbData.dbFullPath);
+//--------------------------------
+// DATABASE OPEN FUNCTION
+//--------------------------------
+
+/** Try to open the database file and return the `DB` handle to it
+ * @param {string} dbLocation the path to the database file
+ * @returns {DB | undefined} the handle to the open database or `undefined` on failure
+ */
+function openDB(dbLocation: string): DB | undefined {
+  if (isString(dbLocation)) {
+    return new DB(dbLocation);
   }
 }
 
-//--------------------------------
-// DISPLAY INFO FUNCTIONS
-//--------------------------------
+//-----------------------------------
+// DISPLAY ACRONYM RECORD FUNCTIONS
+//-----------------------------------
 
 /** Display application version information when requested */
-async function printVersionInfo() {
+async function printVersionInfo(): Promise<void> {
   const versionData = await cliVersion(versionOptions);
   console.log(versionData);
 }
 
-/** Display database data collected when requested */
-function printDbInfo(dbData: DamtInterface) {
+/** Display database file location and stats  */
+async function printDbInfo(dbLocation: string, db: DB): Promise<void> {
   console.log(`
-Database file:      '${dbData.dbFullPath}'
-Database size:      '${dbData.dbSize}'
-Database modified:  '${dbData.dbLastAccess}'
+Database file:      '${normalize(dbLocation)}'
+Database modified:  '${await getFileModTime(dbLocation)}'
+Database size:      '${await getFileSize(dbLocation)}'
 
-SQLite version:     '${dbData.dbSqliteVersion}'
-Total acronyms:     '${dbData.dbRecordCount}'
-Newest acronyms:    '${dbData.dbNewestAcronym}'
-`);
+SQLite version:     '${sqliteVersion(db)}'
+Total acronyms:     '${recordCount(db)}'
+Newest acronym:     '${lastAcronym(db)}'`);
 }
 
 /** Display application help when requested  */
 function printHelp() {
   console.log(`
-Managed acronyms stored in a SQLIte database.
+Managed acronyms stored in a SQLite database.
 
 Usage: ${bold(getAppName())} [switches] [arguments]
 
@@ -269,16 +267,35 @@ Usage: ${bold(getAppName())} [switches] [arguments]
 -l, --latest                        false        show the five newest acronyms records
 -s, --search     acronym            false        acronym to search the database for
 -h, --help                          false        display help information
--v, --version                       false        display program version
-`);
+-v, --version                       false        display program version`);
 }
 
 //--------------------------------
 // Database SQL Functions
 //--------------------------------
 
-//** Search for the given string in the acronym field and display any matching results
-function dbSearch(searchItem: string, db: DB, dbData: DamtInterface) {
+/** Obtain the total number of acronym records in the database */
+function recordCount(db: DB): string {
+  return db
+    ? (db.query("select count(*) from acronyms;")).toLocaleString()
+    : "ERROR";
+}
+
+/** Obtain the SQLite version */
+function sqliteVersion(db: DB): string {
+  return db ? (db.query("select sqlite_version();")).toString() : "ERROR";
+}
+
+/** Obtain the last acronym (ie newest) entered into the database */
+function lastAcronym(db: DB): string {
+  return db
+    ? (db.query("select acronym from acronyms order by rowid desc limit 1;"))
+      .toString()
+    : "ERROR";
+}
+
+/** Search for the given string in the database `acronym` field and display any matching results */
+function dbSearch(searchItem: string, db: DB) {
   if (isString(searchItem)) {
     const searchQuery = db.prepareQuery<
       [string, string, string, string, string, number]
@@ -302,14 +319,16 @@ DESCRIPTION: ${description}`);
       resultCount++;
     }
     console.log(
-      `\nSearch of '${dbData.dbRecordCount}' records for '${searchItem}' found '${resultCount.toLocaleString()}' matches.`,
+      `\nSearch of '${
+        recordCount(db)
+      }' records for '${searchItem}' found '${resultCount.toLocaleString()}' matches.`,
     );
     searchQuery.finalize();
   }
 }
 
-//** Show the last five new acronym records entered into the SQLite database
-function dbLatestRecords(db: DB, dbData: DamtInterface) {
+/** Show the last five new acronym records entered into the SQLite database */
+function dbLatestRecords(db: DB) {
   const latestRecordsQuery = db.prepareQuery<
     [string, string, string, string, string, number]
   >(
@@ -331,83 +350,49 @@ DESCRIPTION: ${description}`);
     resultCount++;
   }
   console.log(
-    `\nShowing latest '5' records of '${dbData.dbRecordCount}' records.`,
+    `\nShowing latest '${resultCount}' records of '${
+      recordCount(db)
+    }' total records.`,
   );
   latestRecordsQuery.finalize();
 }
-
-//** Obtain the total number of acronym records in the database
-function recordCount(db: DB): string {
-  return db
-    ? (db.query("select count(*) from acronyms;")).toLocaleString()
-    : "ERROR";
-}
-
-//** Obtain the SQLite version
-function sqliteVersion(db: DB): string {
-  return db ? (db.query("select sqlite_version();")).toString() : "ERROR";
-}
-
-//** Obtain the last acronym (ie newest) entered into the database
-function lastAcronym(db: DB): string {
-  return db
-    ? (db.query("select acronym from acronyms order by rowid desc limit 1;"))
-      .toString()
-    : "ERROR";
-}
-
-//** Complete the population of the remaining data values in the DamtInterface
-function dbDataPopulate(db: DB, dbData: DamtInterface) {
-  if (db) {
-    dbData.dbSqliteVersion = sqliteVersion(db);
-    dbData.dbNewestAcronym = lastAcronym(db);
-    dbData.dbRecordCount = recordCount(db);
-  }
-}
-
-// sqlite3_prepare_v2(amtdb->db,
-// "select rowid,ifnull(Acronym,''), "
-// "ifnull(Definition,''), "
-// "ifnull(Source,''), "
-// "ifnull(Description,'') "
-// "from ACRONYMS "
-// "Order by rowid DESC LIMIT 5;",
-
-//select rowid,ifnull(Acronym,''),ifnull(Definition,''),ifnull(Source,''), ifnull(Description,''),ifnull(Changed,'') from ACRONYMS Order by rowid DESC LIMIT 5;
 
 //----------------------------------------------------------------
 // MAIN : Deno script execution start
 //----------------------------------------------------------------
 
 if (import.meta.main) {
-  const dbData = await setDbData().catch((e) =>
-    console.error(`Exit as database file not found:\n${e}`)
+  // locate the SQlite database on the filesystem and get its full path
+  const dbLocation = await findDbFileLocation().catch((e) =>
+    console.error(`${e}`)
   );
-  if (!isObject(dbData)) {
-    console.error("ERROR creating database data object");
-  } else {
-    //console.table(dbData);
-    const db = openDB(dbData);
-    if (db) {
-      dbDataPopulate(db, dbData);
-      //console.table(dbData);
-    }
-    if (db) {
-      dbDataPopulate(db, dbData);
+  // check to ensure the path was located - otherwise exit.
+  if (!dbLocation) {
+    console.error("\n[!] FATAL ERROR: database not found. Exit.");
+    Deno.exit(1);
+  }
+  // DEBUG OUTPUT:
+  //console.log(`Found database file: ${dbLocation}`);
 
-      // only returns if execCliArgs() did not find options to execute
-      if (Deno.args.length > 0) await execCliArgs(db, dbData);
+  // Open the database file to obtain `db` handle to the SQLite database
+  const db = openDB(dbLocation);
 
-      // Check if search is needed
-      if (Deno.args.length === 1) {
-        console.log(`Trying search with: '${Deno.args[0]}'`);
-        dbSearch(Deno.args[0], db, dbData);
-      } else {
-        await printVersionInfo();
-        printDbInfo(dbData);
-        printHelp();
-      }
+  // Using the open `db` connection execute the required function.
+  if (db) {
+    // Check for any command line arguments:
+    // NB: only returns if `execCliArgs()` call does not find options to execute.
+    if (Deno.args.length > 0) await execCliArgs(db);
+
+    // Check if search is needed but cli args `-s \ --search` were not provided.
+    if (Deno.args.length === 1) {
+      console.log(`Trying search with: '${Deno.args[0]}'`);
+      dbSearch(Deno.args[0], db);
+    } else {
+      await printVersionInfo();
+      await printDbInfo(dbLocation, db);
+      printHelp();
     }
+
     // close the database
     if (db) {
       db.close();